backend/api_gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = [
    {name = "IvanZaycev0717",email = "ivzaycev0717@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "grpcio (>=1.75.1,<2.0.0)",
    "grpcio-tools (>=1.75.1,<2.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "pydantic[email] (>=2.12.0,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "confluent-kafka (>=2.12.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.1"

